Some very minor changes...
librarian.php
> Added view_profile function, and plans for checkout_reference, which
will decrement or increment the total available of a reference, and an
unnamed function that displays all reserved, borrowed, and waitlisted
references together with the people involved and start and deadline of
involvement.
home_view.php
> Added link to view profile for the librarian
header.php
> Same as home_view.php
uploadSuccess_view.php
> Modified attribute of some tags to set their value based on the value
of the data of the file uploaded.